# Improve checking of unit items