--- a/addons/common/XEH_PREP.hpp
+++ b/addons/common/XEH_PREP.hpp
@@ -178,6 +178,7 @@ PREP(toHex);
 PREP(toNumber);
 PREP(unhideUnit);
 PREP(uniqueElements);
+PREP(uniqueItems);
 PREP(unloadPerson);
 PREP(unloadPersonLocal);
 PREP(unmuteUnit);