--- a/addons/common/functions/fnc_hasItem.sqf
+++ b/addons/common/functions/fnc_hasItem.sqf
@@ -18,4 +18,4 @@
 
 params [["_unit", objNull, [objNull]], ["_item", "", [""]]];
 
-_item in items _unit
+_item in (_unit call EFUNC(common,uniqueItems))