--- a/addons/common/functions/fnc_useItem.sqf
+++ b/addons/common/functions/fnc_useItem.sqf
@@ -22,7 +22,7 @@ private _return = false;
 
 if !(_vehicleUsage) then {
     if (_item != "") then {
-        if (_item in items _unit) then {
+        if (_item in (_unit call EFUNC(common,uniqueItems))) then {
             _unit removeItem _item;
             _return = true;
         } else {