--- a/addons/sandbag/functions/fnc_canDeploy.sqf
+++ b/addons/sandbag/functions/fnc_canDeploy.sqf
@@ -17,6 +17,6 @@
 
 params ["_unit"];
 
-if !("ACE_Sandbag_empty" in items _unit) exitWith {false};
+if !("ACE_Sandbag_empty" in (_unit call EFUNC(common,uniqueItems))) exitWith {false};
 
 _unit call EFUNC(common,canDig)