--- a/addons/trenches/functions/fnc_canContinueDiggingTrench.sqf
+++ b/addons/trenches/functions/fnc_canContinueDiggingTrench.sqf
@@ -18,7 +18,7 @@
 
 params ["_trench", "_unit"];
 
-if !("ACE_EntrenchingTool" in items _unit) exitWith {false};
+if !("ACE_EntrenchingTool" in (_unit call EFUNC(common,uniqueItems))) exitWith {false};
 if ((_trench getVariable [QGVAR(progress), 0]) >= 1) exitWith {false};
 
 // Prevent removing/digging trench by more than one person