--- a/addons/vehiclelock/functions/fnc_hasKeyForVehicle.sqf
+++ b/addons/vehiclelock/functions/fnc_hasKeyForVehicle.sqf
@@ -25,11 +25,12 @@ if (isNull _veh) exitWith {ERROR("null vehicle"); false};
 private _returnValue = false;
 
 //Master can open anything "no matter what"
-if ("ACE_key_master" in (items _unit)) then {_returnValue = true};
+private _items = _unit call EFUNC(common,uniqueItems);
+if ("ACE_key_master" in _items) then {_returnValue = true};
 
 //Check side key
 private _sideKeyName = [_veh] call FUNC(getVehicleSideKey);
-if (_sideKeyName in (items _unit)) then {_returnValue = true};
+if (_sideKeyName in _items) then {_returnValue = true};
 
 //Check custom keys
 private _customKeys = _veh getVariable [QGVAR(customKeys), []];