**When merged this pull request will:**
- Add `uniqueItems` common function - same concept as `CBA_fnc_uniqueUnitItems` however, caches items if unit is `ACE_player` (improves performance when checking units with lots of items)
- Optimize `ace_dogtags_fnc_addDogtagActions` by using `uniqueItems`
- Optimize `ace_explosives_fnc_getDetonators` by using `uniqueItems`

Performance differences for `ace_dogtags_fnc_addDogtagActions`:

Items | Old | New
--- | --- | ---
10 | 0.0683 ms | 0.0826 ms
25 | 0.122249 ms | 0.082 ms
50 | 0.212993 ms | 0.0867 ms
75 | 0.308928 ms | 0.0887 ms
100 | 0.400641 ms | 0.0924 ms
300 | 1.13636 ms | 0.0978 ms

Performance differences for `ace_explosives_fnc_getDetonators`:

Items | Old | New
--- | --- | ---
5 | 0.0532 ms | 0.0945 ms
50 | 0.189609 ms | 0.107308 ms
100 | 0.335233 ms | 0.123047 ms
200 | 0.626959 ms | 0.132293 ms