@mharis001 2018-05-17
Based on slack discussion with @dedmen , unsure whether checking `"my_item" in items _unit` should be changed as well.