@dedmen 2018-05-18
>unsure whether checking "my_item" in items _unit should be changed as well. 

replacing `items` has a noticeably positive effect once you go over about 100 item's I'd say. Question is just if we think that the average user carries more than 100 items.