@Monsen007 2018-08-30
I'm sorry but speak no English ..: / only German .. can you tell me somehow where I have to replace it so where exactly? would be helpful because I do not see it at the janzen ace because it is the steam version


ps translator: P
________________________________
Von: mharis001 <notifications@github.com>
Gesendet: Donnerstag, 30. August 2018 03:19:04
An: acemod/ACE3
Cc: Subscribed
Betreff: Re: [acemod/ACE3] Improve checking of unit items (#6350)


Replacing the cachedCall part (0.0118 ms):

[_unit, _fnc_getItems, _unit, QGVAR(uniqueItemsCache), ITEMS_CACHE_TIME, "cba_events_loadoutEvent"] call FUNC(cachedCall);

with something like this (0.0065 ms):

private _items = _unit getVariable QGVAR(uniqueItemsCache);
if (isNil "_items") then {
    _items = call _fnc_getItems;
    _unit setVariable [QGVAR(uniqueItemsCache), _items];
};
+_items

and a loadout event added in postInit to clear the cache seems worth it, considering we are going for overkill.

This was tested with a low item count (4 unique items), the situation in which the items (~0.0013 ms) command is faster than this solution. This change would help decrease the difference.

And for similar reasons, I think _fnc_getItems should be used rather than CBA_fnc_uniqueUnitItems.

—
You are receiving this because you are subscribed to this thread.
Reply to this email directly, view it on GitHub<https://github.com/acemod/ACE3/pull/6350#issuecomment-417158223>, or mute the thread<https://github.com/notifications/unsubscribe-auth/Ao1sBpWTnAUeQCode-6iHY79nM_OaJUJks5uVz2IgaJpZM4UD1UM>.