@dedmen 2018-08-30
Keep in mind that loadout event only works for local player. the `_unit` there makes it look like it works everywhere