@mharis001 2018-09-16
With no caching at all (4 items): `_fnc_getItems` (0.0085 ms) vs `[_unit, false, true, true, true, false] call CBA_fnc_uniqueUnitItems` (0.0241 ms), about 3x faster.

I think the `uniqueItems` function is as optimized as it will get now, just have to go through and replace any `items _player` that are left.