@mharis001 2018-09-16
I think this PR is complete (replaced most/all `items` command uses), ready for review.